Added functionality to check for non-aplhanumeric characters when a user enters their username. Added the question index feature to show for the questions. Added the confirmation message when a user presses on a button either to proceed to another page or to submit or skip an answer etc. Made the answer input field border color to change when active

+++ js/app.js
@@ -12,13 +12,15 @@ fetch("https://codecyprus.org/th/api/list").then(response => response.json())
     // console.log(element);
     let valid = (element.startsOn <= cD) ?  true : false;
     const button = document.createElement('button');
-    button.innerHTML = element.name;
+
     if(!valid) {
         button.classList.add('invalid');
-        button.disabled = true;
+        button.innerHTML = element.name + " (Not available yet)";
     }
     else {
         button.classList.add('valid');
+        button.innerHTML = element.name;
+        //button.addEventListener('onclick',buttonConfirmation);
     }
     button.classList.add('alertbutton');
    container.appendChild(button);
@@ -27,11 +29,14 @@ fetch("https://codecyprus.org/th/api/list").then(response => response.json())
       button.addEventListener('click', function() {
         let huntName = button.textContent;
         console.log(huntName);
-        if(valid) alert("Starting treasure hunt: " + huntName);
-        window.location.href = "start.html?treasure-hunt-id=" + element.uuid;
-        localStorage.setItem("timeOfTH", element.maxDuration);
+        if(valid) {
+            let confirm  = window.confirm("Are you sure you want to proceed to the '" +  huntName + "' ?");
+            if(confirm){
+                window.location.href = "start.html?treasure-hunt-id=" + element.uuid;
+                localStorage.setItem("timeOfTH", element.maxDuration);
+            }
+        }
       });
 });
 });
-
